Create README.md
Your task is to write a function that takes two arguments, a string and an integer width.  The function returns the string, but with line breaks inserted at just the right places to make sure that no line is longer than the column number. You try to break lines at word boundaries.  Like a word processor, break the line by replacing the last space in a line with a newline.